Factored out friendly-frame-util + minor fixes.
